Move shared example tests to incrementalmerkletree::testing
